refactor: Update blockchain routes to use controller functions
